changed: silence warning.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport/XBMC@16956 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/utils/GUIInfoManager.cpp
@@ -1961,7 +1961,7 @@ bool CGUIInfoManager::GetMultiInfoBool(const GUIInfo &info, DWORD dwContextWindo
             else if (compare.Right(6).Equals(",right"))
             {
               compare = compare.Mid(0,compare.size()-6);
-              bReturn = label.Find(compare) == label.size()-compare.size();
+              bReturn = label.Find(compare) == (int)(label.size()-compare.size());
             }
             else
               bReturn = label.Find(compare) > -1;
